Refactor tests, improve CLI menu, and update compiler flags.
Replaced `int` with `size_t` in tests for safer iteration and fixed a linter warning in `VerifyPassword`. Improved CLI menu by moving the explicit exit option and updating output formatting. Enabled stricter compiler warnings for better code quality.

+++ generator/src/Generator.cpp
@@ -27,7 +27,7 @@ std::string Generator::PasswordGenerator::GenerateSimplePassword(bool intelligib
     {
         std::uniform_int_distribution<int> dist(33, 126); // ASCII full range
 
-        for ( int i = 0; i < policy.passwordLength; i++)
+        for (size_t i = 0; i < policy.passwordLength; i++)
         {
             char c = (char)dist(rng);
             while (policy.excludedCharacters.find(c) != std::string::npos)
